requirements.txt
python-telegram-bot==20.7
mercadopago==2.2.1
python-dotenv==1.0.0
apscheduler==3.10.4
sqlalchemy==2.0.23
pytz==2023.3
requests==2.31.0
